LayoutTests/js/typedarray-lastIndexOf-expected.txt

This test checks the behavior of the TypedArray.prototype.lastIndexOf function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.lastIndexOf.length is 1
PASS Int32Array.prototype.lastIndexOf.name is 'lastIndexOf'
PASS isSameFunctionForEachTypedArrayPrototype('lastIndexOf') is true
PASS testPrototypeReceivesArray('lastIndexOf', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

PASS testPrototypeFunction('lastIndexOf', '(2, -500)', array, -1) is true
PASS testPrototypeFunction('lastIndexOf', '(9, 500)', array, 2) is true
PASS testPrototypeFunction('lastIndexOf', '(2)', array, 3) is true
PASS testPrototypeFunction('lastIndexOf', '(5)', array, 1) is true
PASS testPrototypeFunction('lastIndexOf', '(7)', array, -1) is true
PASS testPrototypeFunction('lastIndexOf', '(2, 3)', array, 3) is true
PASS testPrototypeFunction('lastIndexOf', '(2, 2)', array, 0) is true
PASS testPrototypeFunction('lastIndexOf', '(2, 0)', array, 0) is true
PASS testPrototypeFunction('lastIndexOf', '(2, -1)', array, 3) is true
PASS testPrototypeFunction('lastIndexOf', '(2, -2)', array, 0) is true

PASS successfullyParsed is true

TEST COMPLETE
